requirements.txt
config==0.5.1
crewai==0.193.2
astrapy>=1.0.0
fastapi==0.128.0
langchain_community==0.4.1
langchain_huggingface==1.2.0
nltk==3.9.1
pydantic==2.12.5
python-dotenv==1.2.1
sentence_transformers==5.2.0
